requirements.txt
tensorflow
numpy>=1.26.0
librosa>=0.10.1
numba>=0.60.0
llvmlite>=0.43.0
soundfile
edge-tts
pydub
sounddevice
matplotlib
google-cloud-texttospeech
grpcio
edge-tts
pydub
sounddevice
matplotlib
google-cloud-texttospeech
grpcio